refactore x/uibc
