+++ x/uibc/quota/grpc_query.go
@@ -1,4 +1,4 @@
-package keeper
+package quota
 
 import (
 	context "context"
