+++ x/uibc/quota/migrations.go
@@ -1,4 +1,4 @@
-package keeper
+package quota
 
 import sdk "github.com/cosmos/cosmos-sdk/types"
