+++ x/uibc/quota/msg_server.go
@@ -1,4 +1,4 @@
-package keeper
+package quota
 
 import (
 	"context"
