+++ x/uibc/quota/params.go
@@ -1,4 +1,4 @@
-package keeper
+package quota
 
 import (
 	"errors"
