+++ x/uibc/quota/quota.go
@@ -1,4 +1,4 @@
-package keeper
+package quota
 
 import (
 	"strings"
