+++ x/uibc/quota/unit_test.go
@@ -1,4 +1,4 @@
-package keeper
+package quota
 
 import (
 	"testing"
